+++ app/src/components/forms/create-nested-competency.spec.ts
@@ -8,7 +8,7 @@ import { useRouter } from 'vue-router';
 import { apolloClient } from '~/apollo';
 
 import { formkit } from '../../formkit';
-import CreateCompetencyForm from './create-competency-form.vue';
+import CreateNestedCompetency from './create-nested-competency.vue';
 
 const mockUseRouter = useRouter as unknown as MockedFunction<typeof useRouter>;
 
@@ -26,7 +26,7 @@ test('title is required', async () => {
     push,
   }));
   const user = userEvent.setup();
-  render(CreateCompetencyForm, {
+  render(CreateNestedCompetency, {
     global: {
       plugins: [
         createI18n({ legacy: false, fallbackWarn: false, missingWarn: false }),
@@ -52,7 +52,7 @@ test('create competency submit', async () => {
     push,
   }));
   const user = userEvent.setup();
-  render(CreateCompetencyForm, {
+  render(CreateNestedCompetency, {
     global: {
       plugins: [
         createI18n({ legacy: false, fallbackWarn: false, missingWarn: false }),
